fix rx continuous mode, add DIO2ASSWITCHCTRL, diable invertIQ

+++ src/lib/SX126xDriver/SX126x.h
@@ -14,7 +14,7 @@ public:
 
 
     ///////////Radio Variables////////
-    uint16_t timeout;
+    uint32_t timeout;
 
     ///////////////////////////////////
